# project_docs/generated/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CreateOcrDocumentData": {
        "properties": {
          "ocr_document_id": {
            "title": "Ocr Document Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "ocr_document_id",
          "status"
        ],
        "title": "CreateOcrDocumentData",
        "type": "object"
      },
      "CreateOcrDocumentRequest": {
        "properties": {
          "document_type": {
            "title": "Document Type",
            "type": "string"
          },
          "media_id": {
            "title": "Media Id",
            "type": "string"
          }
        },
        "required": [
          "media_id",
          "document_type"
        ],
        "title": "CreateOcrDocumentRequest",
        "type": "object"
      },
...
      "DataEnvelope_CreateOcrDocumentData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateOcrDocumentData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[CreateOcrDocumentData]",
        "type": "object"
      },
...
      "DataEnvelope_OcrDocumentData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OcrDocumentData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[OcrDocumentData]",
        "type": "object"
      },
...
      "DataEnvelope_list_SessionStreamEventEnvelope__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SessionStreamEventEnvelope"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[list[SessionStreamEventEnvelope]]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OcrDocumentData": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_type": {
            "title": "Document Type",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "low_confidence_fields": {
            "items": {
              "type": "string"
            },
            "title": "Low Confidence Fields",
            "type": "array"
          },
          "media_id": {
            "title": "Media Id",
            "type": "string"
          },
          "ocr_document_id": {
            "title": "Ocr Document Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "ocr_document_id",
          "media_id",
          "document_type",
          "status",
          "fields",
          "low_confidence_fields",
          "created_at",
          "updated_at"
        ],
        "title": "OcrDocumentData",
        "type": "object"
      },
...
      "SessionStreamEventEnvelope": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "task_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Run Id"
          }
        },
        "required": [
          "event_id",
          "seq",
          "event_type",
          "session_id",
          "occurred_at"
        ],
        "title": "SessionStreamEventEnvelope",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Store Manager Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ocr-documents": {
      "post": {
        "operationId": "post_create_ocr_document_api_v1_ocr_documents_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOcrDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CreateOcrDocumentData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Create Ocr Document",
        "tags": [
          "ocr-documents"
        ]
      }
    },
    "/api/v1/ocr-documents/{ocr_document_id}": {
      "get": {
        "operationId": "get_ocr_document_detail_api_v1_ocr_documents__ocr_document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ocr_document_id",
            "required": true,
            "schema": {
              "title": "Ocr Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_OcrDocumentData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ocr Document Detail",
        "tags": [
          "ocr-documents"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/stream-events": {
      "get": {
        "operationId": "list_stream_events_api_v1_sessions__session_id__stream_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_list_SessionStreamEventEnvelope__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Stream Events",
        "tags": [
          "session-stream"
        ]
      }
    },
...
  }
}